refactor(widget):优化频谱图绘制组件的数据类型和性能
- 将 CanvasPoint 的坐标类型从 Double 改为 Float 以提高性能
- 更新 CircularGradientStripView 中参数类型从 Double 到 Float
- 修复 draw 方法中的参数注释描述
- 优化数值转换减少不必要的 toFloat() 调用
- 修正最大高度计算逻辑使用 absoluteValue
-为 GradientStripView 添加完整的渐变条绘制功能
- 实现更精确的竖条高度计算和绘制路径
- 添加适当的画布绘制和刷新机制

+++ app/src/main/java/com/example/android/model/CanvasPoint.kt
@@ -1,3 +1,3 @@
 package com.example.android.model
 
-data class CanvasPoint(val x: Double, val y: Double)
+data class CanvasPoint(val x: Float, val y: Float)
